First Wave of Patches
WARNING: This commit will break Margarine! There are still many commands to rewrite and 99% of all commands will break due to an error. This is solely for backup purposes.

- Usersearch function is rebuilt with fancy object sendback, id and nickname search, new items, and the beginning of callback removal.

Transactions function reworked for no more client.

Got rid of a folder for json files. Much easier to see a small list than a list of folders and guess.