Cleanup + Final Migration
All code is now migrated!

Cleanedup several functions and commands

AniList and MAL usernames are now able to be removed through those commands.

Added en-CA to language to fix an issue of changing the language settings and not being able to revert back.

Added a revoke command to delete all user data for a user. Added a cooldown check to prevent spam daily -> revoke commands.

Broke the award command into two commands, one for the general user and one for awarding.