Fixed and Updated userinfo Command
Fixed an issue with userinfo not being able to find a 'name' key. Updated the userinfo command to display the last time the user sent a message on the server and their guild nickname as well. Bot user field has been converted into a user type and will display either bot or human if the user is a bot or not. ID is also no longer displayed in the title with the account tag.

Added a new utility: dateDisplay. Sending a date to the method will return a human-readable string such as 'November 19, 2020'. Date display options are editable in the settings.json file and follow the options used in toLocaleString method for Date types.

Updated the TOS link in about command to utilize the repository's wiki pages rather than a file within the codebase.

Added a setDefaultChannel command in the Config section to allow admins to determine where general bot messages should go (ie: moderation logs when no mod-log is set.)