Bug Fixes For Music Module
Fixed an issue with queueadd using both an error message and a success message for bad URLs.

Fixed several issues in the remove command. Fixed a bug with removing the first song on the list while Margarine is not playing music. Fixed a bug with specifying a higher index than the songs list. Patched a potential bug when removing the first song on the list while not playing.

Updated Music Module based on repository linting rules.